+++ lib/spior/options.rb
@@ -3,12 +3,13 @@ require_relative 'status'
 
 module Spior
   class Options
-    attr_reader :install , :copy, :mac , :interface
+    attr_reader :install , :copy, :mac , :interface , :tor
 
     def initialize(argv)
       @install = false
       @copy = false
       @mac = false
+      @tor = false
       parse(argv)
     end
 
@@ -25,12 +26,16 @@ module Spior
           @copy = true
         end
 
+        opts.on("-c", "--card NAME", "The name of the target network card") do |net|
+          @interface = net
+        end
+
         opts.on("-m", "--mac", "Change your mac") do
           @mac = true
         end
 
-        opts.on("-c NAME", "--card NAME", "Network card to change, optionnal") do |net|
-          @interface = net
+        opts.on("-t", "--tor", "Redirect traffic through TOR") do
+          @tor = true
         end
 
         opts.on("-s", "--status", "Look infos about your current ip") do