+++ lib/spior/runner.rb
@@ -2,12 +2,15 @@ require_relative 'options'
 require_relative 'install'
 require_relative 'copy'
 require_relative 'mac'
+require_relative 'iptables'
+require_relative 'network'
 require_relative 'msg'
 
 module Spior
   class Runner
     def initialize(argv)
       @options = Options.new(argv)
+      @network = false
     end
 
     def run
@@ -21,7 +24,17 @@ module Spior
       end
       if @options.mac then
         Msg.head
-        Spior::MAC::randomize(@options.interface)
+        if not @network
+          @network = Spior::Network.new(@options.interface)
+        end
+        Spior::MAC::randomize(@network.card)
+      end
+      if @options.tor then
+        Msg.head
+        if not @network
+          @network = Spior::Network.new(@options.interface)
+        end
+        Spior::Iptables::tor(@network.card)
       end
     end
   end