+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.5, release 2020-05-03
+* Spior can now redirect all the traffic through TOR
+* Add OptionParser -t|--tor
+* Change class lib/spior/mac by module
+* Correct option -c|--card NAME
+* Create lib/spior/network
+* Create lib/spior/iptables 
+
 ## 0.0.4, release 2020-05-02
 * Spior can now randomize a mac address with decertmac
 * Add field s.metadata to spior.gemspec